drop wal files on failed replay
