+++ tempodb/wal/wal.go
@@ -5,7 +5,6 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strings"
 
 	"github.com/google/uuid"
 )
@@ -17,13 +16,6 @@ const (
 // WAL allows for interaction with the WAL
 type WAL struct {
 	c *Config
-}
-
-// File represents a walfile entry
-type File struct {
-	Filepath string
-	TenantID string
-	BlockID  uuid.UUID
 }
 
 type Config struct {
@@ -108,22 +100,4 @@ func (w *WAL) NewBlock(id uuid.UUID, tenantID string) (*AppendBlock, error) {
 // NewBlockWithWalFile creates an AppendBlock preloaded with contents of a WalFile
 func (w *WAL) NewBlockWithWalFile(f *File) (*AppendBlock, error) {
 	return newAppendBlockFromWal(f)
-}
-
-func parseFilename(name string) (uuid.UUID, string, error) {
-	i := strings.Index(name, ":")
-
-	if i < 0 {
-		return uuid.UUID{}, "", fmt.Errorf("unable to parse %s", name)
-	}
-
-	blockIDString := name[:i]
-	tenantID := name[i+1:]
-
-	blockID, err := uuid.Parse(blockIDString)
-	if err != nil {
-		return uuid.UUID{}, "", err
-	}
-
-	return blockID, tenantID, nil
 }
